Split settings into StaticSettings and SettingsService.
The first is the format SettingsServer the last (new one) will be used for the admin settings.
