+++ server/src/database/models/settings.model.ts
@@ -1,0 +1,5 @@
+import { modelOptions } from '@typegoose/typegoose';
+import { CollectionName } from '../../helpers/CollectionName';
+
+@modelOptions({ schemaOptions: { collection: CollectionName.SETTINGS } })
+export class SettingsModel {}
