+++ server/src/module/settings/settings.module.ts
@@ -1,4 +1,4 @@
-import { Module, Global } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
 import { SettingsService } from './settings.service';
 
 @Global()
