+++ server/src/module/settings/settings.service.spec.ts
@@ -1,15 +1,26 @@
 import { Test, TestingModule } from '@nestjs/testing';
+import { TestModule } from '../../../test/helpers/test.module';
 import { SettingsService } from './settings.service';
 
 describe('SettingsService', () => {
+  let testModule: TestingModule;
   let service: SettingsService;
 
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
+  beforeAll(async () => {
+    testModule = await Test.createTestingModule({
+      imports: [TestModule.forRootAsync()],
       providers: [SettingsService],
     }).compile();
+  });
 
-    service = module.get<SettingsService>(SettingsService);
+  afterAll(async () => {
+    await testModule.close();
+  });
+
+  beforeEach(async () => {
+    await testModule.get<TestModule>(TestModule).reset();
+
+    service = testModule.get<SettingsService>(SettingsService);
   });
 
   it('should be defined', () => {
